Modifications for it to build on OSX

+++ libiberty/regex.c
@@ -1048,7 +1048,7 @@ PREFIX(print_partial_compiled_pattern) (UCHAR_T *start, UCHAR_T *end)
           PREFIX(extract_number_and_incr) (&mcnt, &p);
 	  p1 = p + mcnt;
           PREFIX(extract_number_and_incr) (&mcnt2, &p);
-	  printf ("/jump_n to %ld, %d times", p1 - start, mcnt2);
+	  printf ("/jump_n to %ld, %d times", (long int)(p1 - start), mcnt2);
           break;
 
         case set_number_at:
